HBX-1550: Inline class 'org.hibernate.cfg.reveng.SimpleMetaAttribute' into class 'org.hibernate.cfg.reveng.MetaAttributeBinder'
