--- main/src/main/java/org/hibernate/cfg/reveng/SimpleMetaAttribute.java
@@ -1,17 +1,0 @@
-package org.hibernate.cfg.reveng;
-
-public class SimpleMetaAttribute {
-	String value;
-	boolean inheritable = true;
-
-	public SimpleMetaAttribute(String value, boolean inherit) {
-		this.value = value;
-		this.inheritable = inherit;
-	}
-
-	public String toString() {
-		return value;
-	}
-
-}
-
